Add: Echo-Contacts API
Echo-Contacts is repository that demonstrates the use of `echo` framework in creating contacts.

The repository uses `mysql` as db driver and the below given endpoints can be used to `Create` , `Read` , `Update` and `Delete`.

The endpoints uses jwt token based authorization and the token can be generated using the `/login` endpoint.